+++ sample-apps/react/react-dogfood/components/Lobby.tsx
@@ -53,25 +53,31 @@ export const Lobby = ({ onJoin, mode = 'regular' }: LobbyProps) => {
     useCallSession,
     useCallMembers,
     useCallCustomData,
+    useCallSettings,
   } = useCallStateHooks();
   const { hasBrowserPermission: hasMicPermission } = useMicrophoneState();
   const { hasBrowserPermission: hasCameraPermission, isMute: isCameraMute } =
     useCameraState();
   const callSession = useCallSession();
   const members = useCallMembers();
+  const settings = useCallSettings();
   const currentUser = useConnectedUser();
   const isProntoEnvironment = useIsProntoEnvironment();
   const isDemoEnvironment = useIsDemoEnvironment();
   const [displayNameOverride, setDisplayNameOverride] = useState<string | null>(
     isDemoEnvironment ? getRandomName() : null,
   );
-  const displayName = displayNameOverride ?? currentUser?.name ?? '';
+  const router = useRouter();
+  const displayName =
+    displayNameOverride ??
+    currentUser?.name ??
+    (router.query['user_id'] as string | undefined) ??
+    '';
   const custom = useCallCustomData();
 
   const { t } = useI18n();
   const edges = useEdges();
 
-  const router = useRouter();
   const skipLobby =
     !!router.query['skip_lobby'] ||
     process.env.NEXT_PUBLIC_SKIP_LOBBY === 'true';
@@ -145,22 +151,28 @@ export const Lobby = ({ onJoin, mode = 'regular' }: LobbyProps) => {
                   )}
                 >
                   <div className="rd__lobby-video-preview">
-                    <VideoPreview
-                      DisabledVideoPreview={
-                        hasBrowserMediaPermission
-                          ? DisabledVideoPreview
-                          : AllowBrowserPermissions
-                      }
-                    />
+                    {settings?.video.enabled ? (
+                      <VideoPreview
+                        DisabledVideoPreview={
+                          hasBrowserMediaPermission
+                            ? DisabledVideoPreview
+                            : AllowBrowserPermissions
+                        }
+                      />
+                    ) : (
+                      <DisabledVideoPreview />
+                    )}
                     <div className="rd__lobby-media-toggle">
                       <ToggleAudioPreviewButton Menu={null} />
-                      <ToggleVideoPreviewButton Menu={null} />
+                      {settings?.video.enabled && (
+                        <ToggleVideoPreviewButton Menu={null} />
+                      )}
                     </div>
                   </div>
                   <div className="rd__lobby-controls">
                     <div className="rd__lobby-media">
                       <ToggleMicButton />
-                      <ToggleCameraButton />
+                      {settings?.video.enabled && <ToggleCameraButton />}
                     </div>
 
                     <div className="rd__lobby-settings">